+++ apps/autotest/models/TST/_migrate/999999999/010-test_beforeDDL.sql
@@ -1,0 +1,3 @@
+delete from uba_user where ID < 0
+--
+delete from ub_blobHistory where ID < 0